+++ src/module/actor/character-sheet.js
@@ -31,6 +31,9 @@ export class OseActorSheetCharacter extends OseActorSheet {
           initial: "attributes",
         },
       ],
+      scrollY: [
+        '.inventory'
+      ],
     });
   }